% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doSimLast}
\alias{doSimLast}
\title{Do Automata Simulation (F fixed for Ramos & Leite)}
\usage{
doSimLast(AlphaProb, Replication, Size = 1000L, MaxIterations = 10000L,
  InitialProb = 0.5, Neighbors = as.integer(c(-1, 0, 1)))
}
\arguments{
\item{\code{AlphaProb}}{Probability of change}

\item{\code{Replications}}{Number of replications. Default 100}

\item{\code{Size}}{Number of components in the configuration}

\item{\code{MaxIterations}}{Default 1000}

\item{\code{InitialProb}}{Probability for initial configuration}

\item{\code{Neighbors}}{Integer vectors of neighbors. Defaulf c(-1,0,1)}
}
\value{
A data frame with Size, Replication, Iteration, AlphaProb and Frequency of Symbols
}
\description{
Save the space and temporal mean of configurations.
}

