% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_simulation}
\alias{grid_simulation}
\title{Do Automata Simulation (F with Alpha and Beta fixed for Ramos & Leite 2016)}
\usage{
grid_simulation(threads = 1L, step = 0.01, initial_block_lenght = 1L,
  max_iterations = 100000L, line_size = 1000L,
  Neighbors = as.integer(c(-1, 0, 1)))
}
\arguments{
\item{\code{line_size}}{Number of components in the configuration}

\item{\code{Step}}{Alpha and beta increment}

\item{\code{max_iterations}}{Default 1000}

\item{\code{replications}}{Default 1000}

\item{\code{initial_block_size}}{Inital block size. Default 1.}

\item{\code{Neighbors}}{Integer vectors of neighbors. Defaulf c(-1,0,1)}
}
\value{
Data frame. If all zeros, return 0. If all ones, return 1. If MaxIterations reached, return -1.
}
\description{
Save the space and temporal mean of configurations.
}

