% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatrixSample}
\alias{MatrixSample}
\title{Generate a random integer matrix}
\usage{
MatrixSample(n, k, p)
}
\arguments{
\item{\code{n}}{Integer}

\item{\code{k}}{Integer}

\item{\code{p}}{Double in (0,1)}
}
\value{
Random matrix
}
\description{
Generate a random integer matrix
}

