% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simIteration}
\alias{simIteration}
\title{Do a Simulation Step (Only 3 neighbors...)}
\usage{
simIteration(X, Alpha, Neighbors)
}
\arguments{
\item{\code{X}}{Numeric vector representing a configuration}

\item{\code{Alpha}}{Probability of change}

\item{\code{Neighbors}}{Integer vectors of neighbors}
}
\value{
A List with new state and frequency of symbols
}
\description{
Do a Simulation Step (Only 3 neighbors...)
}

