% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{simulation}
\alias{simulation}
\title{Parallel automata simulation}
\usage{
simulation(Replications = 100, MaxIterations = 1000, Alpha, Beta,
  Size = c(1000), InitialProb = c(0.5))
}
\arguments{
\item{\code{Replications}}{Number of replications. Default 100.}

\item{\code{MaxIterations}}{Max number of iterations. Default 1000.}

\item{\code{Alpha}}{Alpha probability vector. Default seq(from = 0, to = 1, by = 0.25).}

\item{\code{Beta}}{Beta probability vector. Default seq(from = 0, to = 1, by = 0.25).}

\item{\code{Size}}{Size vector. Default c(1000).}

\item{\code{InitialProb}}{Initial probability vector. Default c(.5).}
}
\value{
Dataframe.
}
\description{
Parallel automata simulation
}

