\name{toom-package}
\alias{toom-package}
\alias{toom}
\docType{package}
\title{
\packageTitle{toom}
}
\description{
\packageDescription{toom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{toom}
\packageIndices{toom}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{toom}

Maintainer: \packageMaintainer{toom}
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g. 
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
Simple examples of the most important functions
}
