% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VectorSample}
\alias{VectorSample}
\title{Generate a random integer vector}
\usage{
VectorSample(n, p)
}
\arguments{
\item{\code{n}}{Integer}

\item{\code{n}}{Double}
}
\value{
Random vector
}
\description{
Generate a random integer vector
}

